MDL-43749 DO NOT USE! quiz database upgrade code
Here is the database upgrade code for this issue. However, don't use it
because I have only done the database upgrade part. I have not changed
the code that uses this data, so applying this patch will break your
site.

I have only committed and pushed this for these reasons:
* I wanted an off-site backup.
* If anyone would like to review these changes, then I would love to get
the feedback.